+++ CHANGELOG.md
@@ -2,6 +2,9 @@ Changelog
 -------------
 
 ### Unreleased
+* Add user customisation (based on the work of @nieldm [#163](https://github.com/televator-apps/vimari/pull/163)).
+* Update Vimari interface to allow users access to their configuration.
+* Remove `closeTabReverse` action.
 
 * Normal mode now isolates keybindings from the underlying website, this means that to interact with the underlying website you need to enter insert mode.
 * You can enter insert mode by pressing <kbd>i</kbd> and exit the mode by pressing <kbd>esc</kbd>.